Handle current package whose full version is specified in the changelog
Before, the 'list' command didn't show built package whose version does
not contain the local version tag ("beip"), but the build command could
build such a package if the version is already specified in the
changelog.

Now a current "foreign" package will correctly be detected as built in
that case.