Expect we will be ran from the repository directory.
Before, we were expecting a single 'autobuild' directory was there and
the actual datafiles were in that directory. This is a relict from the
old build scripts we used a few years ago.

The behaviour was changed to match more closely what other similar tools
do, i.e. expect the command is ran from the actual repository directory.