Make it possible to provide one's own revision pattern.
The debian revision pattern is now configurable. By default, it is just
'%s' so basically the Build-Number value gets appended to the version
number. This option is not really meant to be given through command
line, but can be added in the novabuild.cfg file.

So basically there is now two ways of working wrt build numbers:

  a/ using a straight number in the moduleset and configuring a pattern
  b/ using fancy build numbers staight into the moduleset