Automatically append "build tags" to the latest changelog entry.
This way we can use the same changelog files for several builds
differing only by a tag.

The purpose is to be able to use the same (templatized) data files for
squeeze and wheezy builds. This is a bit of a bending of debian good
practice rules but it seems to work like a charm.